backend/requirements.txt
flask>=3.0.2
werkzeug>=3.0.0
textblob==0.19.0
nltk==3.9.1
requests==2.31.0
python-dotenv==1.0.1
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
sqlalchemy==2.0.23
pillow==10.1.0
pytesseract==0.3.10
numpy==1.26.2
scikit-learn==1.3.2
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
bcrypt==4.0.1
pyjwt==2.8.0
python-dateutil==2.8.2
email-validator==2.1.0.post1 